refactor: use settings.quiet instead of wandb.jupyter.quiet() (#9705)
Drops the `wandb.jupyter.quiet()` method and the `-q`/`--quiet` option to the `wandb` magic. The only use of this method was in `wandb_run.py` right below a TODO about using settings instead.

This is technically a breaking change, but the option does not appear to have ever been documented anywhere, and "broken" notebooks are trivial to fix.