refactor: remove the workspace option to wandb ipython magic (#9709)
Removes the `-w` / `--workspace` option to the `wandb` cell/line IPython magic because it is a no-op.

I attempted fixing it as well, but it does not appear to work: using `%%wandb -w` in a cell that initializes a run produces an iframe that fails to load.

The original code:

```python
if self.opts.get("workspace", False):
    if self.path is None and wandb.run:
        self.path = wandb.run.path
if isinstance(self.path, str):
    object = self.api.from_path(self.path)
else:
    object = wandb.run
if object is None:
    if wandb.Api().api_key is None:
        return "You must be logged in to render wandb in jupyter, run `wandb.login()`"
    else:
        object = self.api.project(...)
return object.to_html(self.height, hidden=False)
```

If `workspace` is False:

* If `self.path` is set (`%%wandb entity/project/id`), it is used
* Else, if `wandb.run` is set, it is used

If `workspace` is True:

* If `self.path` is set, it is used
* Else, if `wandb.run` is set, its path is used
    * `wandb.run.path` always has 3 components, so `api.from_path()` returns a `public.Run`
    * `public.Run.to_html` is identical to `wandb.run.to_html`

So `workspace` did not have any effect.