refactor: avoid global run in wandb ipython magic (#9710)
Refactors the wandb IPython magic (`%%wandb` / `%wandb`) to avoid using the global run when displaying a run as part of `wandb.init()`.

The old `maybe_display()` function was only called when printing the Run header when initializing a run, but it used `wandb.run` to decide what to display. Now it accepts a run explicitly.

`wandb.run` is still used when running the `%wandb` magic on its own, like

```python
# --- Cell 1 ---
run = wandb.init()

# --- Cell 2 ---
# This creates an iframe displaying the above run's page.
%wandb -h 600
```

On the other hand, this uses the explicit run and will continue to work when `wandb.run` is removed:

```python
%%wandb -h 600
with wandb.init() as run:
    ...
```