refactor: remove wandb-core message from wandb.init() (#9733)
Fixes WB-24478.

Instead of printing a message when the user is on `wandb-core`, which has been the norm for a long time now, print a warning on init when the user is on `legacy-service`.

The new warning is in addition to the warning printed at the end of each `legacy-service` run. Having it at the start can help users (realistically, only developers) notice if they're accidentally using `legacy-service`.

To test:

* `python -c 'import wandb; wandb.init()'`
* `WANDB_X_REQUIRE_LEGACY_SERVICE=true python -c 'import wandb; wandb.init()'`