refactor: rewrite RunParams and clean up fork/rewind/resume.go (#9791)
Cleans up `internal/runbranch` in wandb-core to prepare for an UpsertBucket-related refactor:

* Simplified `RunParams` and organized it for better readability
* Changed fork/rewind/resume.go to modify the `RunParams` in-place instead of returning updates that are then merged
* Instead of accepting `RunParams` and `RunPath`, fork/rewind/resume.go read the entity/project/run ID from `RunParams`. The `RunPath` parameter was always derived from `RunParams` in `Sender`.
* Removed the `Config` field of `RunParams` and made rewind/resume.go accept and mutate a `RunConfig` directly

In the `Sender`, `startState` now starts out as `nil` and becomes set when processing the first `RunRecord` (if an error occurs, it is set back to `nil`).

`Sender` no longer clones the `RunRecord` at the start of `sendRun`: it is cloned immediately before it is mutated (by merging `startState.Proto()` into it) wherever that happens.