refactor: create internal/runmetadata (#9814)
Creates `internal/runmetadata` and fills it out with unimplemented methods to serve as an easy-to-review base for later PRs that implement it in parts.

`RunMetadata` will contain the majority of the implementation.

`RunUpdateWork` is the glue that makes `Handler` and `Sender` interact with `RunMetadata`. The `RunMetadata` object will live on a `Stream`, and `Stream` will turn `RunRecord` into `RunUpdateWork`. Its `Accept()` method returns `true`, corresponding to `Handler` unconditionally forwarding the record; `Process()` is the `Sender` portion. `BypassOfflineMode()` returns true corresponding to `Handler` setting `AlwaysSend` on the `RunRecord`.