refactor: runmetadata UpsertBucket (#9815)
Implements the `UpsertBucket` request in `internal/runmetadata`, except for portions related to the run config which are added in a separate PR.

The key method is `lockedDoUpsert` which is called when initializing a run or flushing changes to its parameters (notes, tags, etc.) or config.

To allow testing, this removes a few `panic()` statements and replaces them with TODO comments.