refactor: paths.LongestCommonPrefixStr and paths.Join, tests (#9858)
Adds two new functions to `paths` and adds some missing unit tests.

The functions are used in later PRs: `LongestCommonPrefixStr` is the core logic of `LongestCommonPrefix` but not necessarily requiring a real system path, which is used when comparing cloud URLs in `internal/tensorboard`.

`paths.Join` is natural to have here, but I didn't add it before because it wasn't needed.