refactor: runmetadata resume, rewind, fork (#9818)
Fills out resuming, rewinding and forking in `internal/runmetadata`.

NOTE: The original code in `sender.go` does not do `UpsertBucket` when rewinding a run (by setting `resume_from`), but does it for resuming and forking. `RunMetadata` does it in all cases.

This PR relocates some `resume` tests to `test_resume.py` and adds a few new ones. There are no rewind/fork system tests because we don't support those in CI, so that's only covered by the Go unit tests.