refactor: enable Ruff UP032 lint (use f-strings instead of format()) (#9897)
Removes `UP032` from the ignore list and applies automatic fixes to all files using `ruff check . --fix`.

As a result, some lines may be a bit overlong, but it's difficult to auto-fix, and enforcing the lint is worth it.

I ignore the lint for `wandb/apis/public/` because f-strings need to be formatted a little differently for `gql(...)` syntax highlighting to work.