refactor: fix a few instances of TRY300 (#9918)
https://docs.astral.sh/ruff/rules/try-consider-else/

This one takes some manual work to really fix, and I don't enable this lint in this PR because its suggestion can be a little misleading. For instance, it will produce a diagnostic on this:

```python
try:
    result = computation()
    return result
except SomeError:
    return other
```

The lint message will make it seem that the solution is this:

```python
try:
    result = computation()
except SomeError:
    return other
else:
    return result
```

But an even better solution is

```python
try:
    return computation()
except SomeError:
    return other
```

In many cases in our codebase, the right fix is to narrow the scope of the `try-except` rather than to use `try-except-else`. It also sometimes makes sense to not use an `else` block at all, especially when the exception handler is an early return.