refactor: use `shutil.which()` instead of find_executable() from dockerpycreds (#9934)
The `find_executable()` in `dockerpycreds` used to wrap `distutils.spawn.find_executable()` to additionally handle `PATH_EXT` on Windows, which is why we used it according to comments in the code. The replacement suggested by [PEP 0632](https://peps.python.org/pep-0632/#migration-advice) is `shutil.which()` which appears to have logic using `PATH_EXT` (at least in Python 3.10).