refactor: enable and fix TRY300 (try-consider-else) (#9923)
https://docs.astral.sh/ruff/rules/try-consider-else/

The lint is a little misleading: it's not about preferring an `else` clause in `try-except` blocks, but about narrowing the scope of a `try-except`. So, for instance:

```python
try:
    result = do_something()
    do_something_else()
    return result
except DoSomethingError:
    return something_else
```

should really be written as

```python
try:
    result = do_something()
except DoSomethingError:
    return something_else

do_something_else()
return result
```

An `else` clause would be appropriate if the `except DoSomethingError` clause didn't return, or if it helps with readability by improving vertical alignment or flow of logic.

I fixed a few cases, and added ignores for the remaining violations. This at least prevents this kind of code in new files.