refactor: rewrite service_token.py to prepare for Unix domain sockets (#9967)
Makes `ServiceToken` an abstract class with `connect()` and `save_to_env()` methods, with the current localhost-socket implementation provided by `NetworkServiceToken`.

Moving the `connect()` responsibility into `service_token.py` allows reusing the same code when connecting to either an environment token or an explicitly constructed token, and this makes it possible to provide an implementation for Unix domain sockets later.