refactor: move wandb/sdk/service/service.py to wandb/sdk/lib/service/ (#9961)
Moves `service.py`, which has to do with starting up the service, into `wandb/sdk/lib/service/` next to `service_connection.py` which establishes a connection and is the only file that references it.

Most of the remaining code in the `wandb/sdk/service/` folder is legacy-service code. I will separate out the read portion of `port_file.py` in a separate PR.