chore: pin Ruff in pre-commit check (#10062)
It seems that sometimes, the `Ruff` version used by `pre-commit` is unstable and can even go back in time. Although we specified a version in `pre-commit-config.yaml`, I think the `pip install` in `pre-commit.yml` sometimes messed with this.

This pins to the latest `Ruff` (0.12.0), applies formatting fixes, and updates some outdated `noqa` directives (needed because `UP007` got split into `UP007` and `UP045`).