fix: clear console capture callbacks on error (#10071)
Makes `console_capture.py` clear all callbacks if a callback raises an error.

Since callbacks are invoked when stdout or stderr is being written to, and since errors are likely to be printed to the console, this can trigger an infinite loop. This is especially problematic for `KeyboardInterrupt` as it makes it difficult to stop the program using Ctrl+C if a callback is permanently stuck (imagine a callback that does `time.sleep(600)`).

Also updates `console_capture.py` to log and suppress regular Exceptions thrown by callbacks, since these will almost always be W&B bugs that shouldn't propagate to the user's `print()` statements.