chore(artifacts): compute hash of multipart in parallel
Hashing is taking longer than upload after enabling multipart upload for GCS.
Both python and go code are hashing file. Python is hashing entire file,
md5 hash is serial by design of the algo. Go is hashing parts of file
independently so we can hash the parts in parallel.

The change in go is backward compatiable. There is no change for python
hash logic in this commit.