refactor: rethink where TensorBoard lives (#10039)
Moves TensorBoard awareness to an earlier stage in wandb-core: the `work.Schedule()` method, which happens before `Accept()` (Handler) and `Process()` (Sender).

## Motivation

This is motivated by the TensorBoard integration creating the last remaining case where the Sender must send messages back to the Handler: on receiving an Exit record, the Sender waits to shut down TensorBoard because it may generate additional records. This PR allows completely getting rid of this mechanism, replacing `Sentinel()` by `Schedule()`. The next PR does this refactor.

## Design notes

The design reasoning for this PR is that the simplest way to treat the TensorBoard integration is to pretend the client is directly calling run methods (`run.log()`, `run.save()`). If this were true, then the client would shut down TensorBoard before sending an Exit record; we simulate this at `Schedule()` time to maintain the illusion.

There's a nuance that will be important later: `Schedule()` can reorder records. For an online run this is OK, since `Schedule()` runs before the Handler when records may be reordered; but during `wandb sync` (not yet implemented in wandb-core), this is problematic because `Schedule()` is given records that have already been written. That's a problem because of flow control (also not yet implemented), for which the simplest implementation assumes records arrive in the same order they appear in the transaction log. The `WaitGroup` parameter to `Schedule()` is the solution: `sync` will `Wait()` on it to prevent reordering.