refactor: no plum-dispatch in metaflow integration (#10202)
Refactors the Metaflow integration to not use `plum-dispatch`, which works poorly with IDEs and static analysis tools (try opening the file in an IDE; supposedly there's a [workaround](https://beartype.github.io/plum/integration.html#overload-support) but I was unable to get it to fully work).

Using `plum-dispatch` it was possible to extend `wandb_track` and `wandb_use` for other types in other packages by using `@wandb_track.dispatch` or `@wandb_use.dispatch`. Neither of these appears on GitHub, and we didn't have this capability documented, so I'm not too worried about removing it.

This PR changes things in place so that it's easier to verify that the functionality didn't change, but I'll move things around in later PRs.