refactor: write to transaction log synchronously (#10060)
Refactors `writer.go` to finish writing records to the transaction log before forwarding them.

This is the original behavior, which we changed because it was unnecessary. It will soon become necessary for "flow control": if we discard Work because the Sender is blocked and then try to read it from the transaction log when it's unblocked, the Work needs to have been fully written.

One of our motivations before was to avoid blocking on the transaction log if writing to disk is slow---but note that blocking would have happened anyway once the finite storeChan buffer filled up.