refactor: fix SeekRecord to allow seeking in the transaction log (#10141)
Makes Store's Reader's `SeekRecord` method work correctly with our customizations to the LevelDB file format, so that it can be used for flow control.

The comment at the top of `record.go` has a good description of the format, but the gist of it is that the stream of data (i.e. the transaction log file) is divided into 32 KiB blocks. The fixed block size makes it possible to read data even if a block is corrupted (skipping any records that had some data in that block).

Our customization keeps the 32 KiB block boundaries, but writes a custom header into the first 7 bytes of the stream. This causes the first block to start at byte 7 and to contain at most (32 * 1024 - 7) bytes. This is regrettable because we could have just put the header into its own LevelDB record keeping the format unchanged.

This PR makes `record.go` aware of the W&B header, replacing the `initialData` Writer parameter and `ReadHeader` Reader method. I decided against extending those abstractions (like by adding a `initialHeaderLength` field to `Reader`) because we only use `record.go` for the transaction log anyway, and mentioning the W&B header explicitly makes the reason for the customization clear. Either way would have worked though.