chore: disable golangci-lint errcheck on std error
We are using the default set of linters from golangci-lint.
Default linters has `errcheck` and checks on `file.Close()`.
This seems to be a new change in either golangci-lint or errcheck.
Use the recommended preset to skip checking `file.Close()` and other
common io closer, which is mostly used in test or defer.
https://golangci-lint.run/usage/false-positives/#preset-std-error-handling

We should consider tweak the config to enable/disable other linters
to align with internal repos golangci-lint config. Though this would
likely requires some clean up of the code base on the new linter config