refactor: create AsyncioManager (#10334)
Creates `AsyncioManager` to allow running `asyncio` in a separate thread and adding tasks to it.

`wandb` currently creates a new `asyncio` thread each time it waits on a mailbox handle, while a separate thread continuously polls a socket and pushes responses to the handles. This is only slightly inefficient (`wandb` programs don't need to wait mailbox handles often), so the main issue is multithreaded use of a socket. The read thread flips the socket between blocking and timeout modes; the write thread is written to work with either, but this is not robust: we've seen it throw a `BlockingIOError` which shouldn't be possible and indicates a potential race.

With `AsyncioManager`, a single long-lived `asyncio` thread will be able to run a socket reading task, occasional socket writing tasks, and mailbox tasks.