refactor: completely replace SockClient and MessageRouter (#10355)
Completely rewrites the socket code used to communicate between the Python client and wandb-core.

The main addition is ServiceClient, and there are many simplifications in Mailbox because all the async operations now happen in a single thread.

The ancient SockClient and MessageRouter classes are replaced by the much smaller and simpler ServiceClient. Instead of using multithreading, all socket communication happens in the asyncio thread. SockClient had a race condition because it flipped the socket between timeout and blocking mode in an unsynchronized way; this is no longer a problem.