chore: add weave integration file (#10403)
Adds a new module `wandb.integration.weave` that exposes a stable interface `weave` can rely on to interact with `wandb`. Not meant for packages other than `weave`.

Also makes a few functions in `wandb_setup.py`​ thread-safe:

- setup(), singleton() and teardown() share a lock, so their invocations always happen in order
- most_recent_active_run(), add_active_run() and remove_active_run() share a lock as well
- finish_all_active_runs() can be called simultaneously with the other `active_run`​ methods but only from one thread at a time (so `wandb.init()`​ is still not thread-safe)
    - This is because `run.finish()`​ is not thread-safe

This way we don’t need to worry if a Run is active in a different thread from `weave.init()`​.