refactor: server request protos for wandb beta sync (#10399)
Adds protos that will be used for communication in `wandb beta sync`. There are three requests:

- `init_sync` generates an ID for the operation, used in the other two requests
- `sync` initiates syncing and responds when done
- `sync_status` polls for status messages

This could be simpler if we had a mechanism to stream multiple responses for a request, in which case a single request could receive zero or more status updates followed by a final update. That's something we can add in the future.