refactor: remove inform_start request (#10449)
Removes it because it has no effect other than to cause confusion.

This was sent by `wandb.init()` after a `Run` record and before a `RunStart` request.

The first thing that `wandb-core` did in response was overwrite the `Stream.settings` field which was too late anyway: everything is initialized by this time to use the first Settings object created using the settings passed to `inform_init`. Even if that worked, the settings don't change between `inform_init` and `inform_start`.

The other thing was to update the "run_url" tag on the Stream's Sentry instance, but again, this is already initialized correctly when the `observability.CoreLogger` is constructed during `inform_init` (see `internal/stream/streamlogger.go::streamLogger`).