refactor: call Printer.progress_close() correctly and remove text param (#10419)
I think the `text` parameter was a remnant of how we used to display `run.finish()`. The fact that this would print an empty line if called with default values was hidden because we didn't call it in a `finally` block.

This PR also fixes a quality issue in Jupyter where some of the `wandb.init()`​ status updates (like “creating run (\<duration>)”) would remain after the operation completed. This was because the code to clear the text was not in a `finally`​ block and was skipped when the asyncio task running it got cancelled.

The `progress_close()`​ method is probably unnecessary: its only effect is to close (hide) a progress bar widget in Jupyter, but we never display one. A progress bar is only displayed when `progress_update()`​ is used, but its only callsite is in `progress.py`​ in a codepath that can’t be reached in Jupyter as far as I can tell.