refactor: filestream request size calculation (#10470)
Cleans up the logic estimating whether the accumulated FileStream request is at maximum size.

`isAtMaxSize` didn't take the summary size into consideration before (`addStringsToRequest`​ is not used for the summary). The only difference this makes is that if the user does something like `run.summary["key"] = "huge value ..."` while turning off system metrics, not logging any data and not printing anything to the console, then we'll now make the request immediately rather than waiting to accumulate more data.

It was odd to use a stateful computation to compute this value.