feat: limit beta sync parallelism (#10469)
Limits how many runs `wandb beta sync` attempts to sync at a time when multiple files are specified, controllable with an `-n` option.

I arbitrarily picked a default of 5. Syncing is IO bound, not CPU bound, so it doesn't make sense to set it to the number of CPUs. It depends on the user's network speed and server load (and custom deployments may have different ideal values).

We can experiment a little bit and tweak this once syncing is fully implemented.