fix: do not infinitely-buffer messages to wandb-core (#10526)
Removes hidden buffering for messages to wandb-core.

To write to the socket, we were previously scheduling asyncio tasks without waiting for them (using `AsyncioManager.run_soon()` in `service_client.py`). There is no limit to how many tasks can pile up, and each task holds on to the data given to it, meaning there is no upper limit to memory usage.

The original reason to use `run_soon()` was that it is safe to call inside either a regular or `async` function, whereas `AsyncioManager.run()` is blocking and so can only be called from a regular function. This PR updates all `InterfaceBase` methods to internally rely on `run()` instead of `run_soon()`, making them illegal to call inside `async` functions, and it also updates all `async` functions that were previously calling those methods.

This is unfortunately hard to lint for. A linter would need to get the call tree of `AsyncioManager.run()`, but I found no easy-to-use library that does this. There's a third party checker "pylint-blocking-calls" but it does not appear to find violations and is very slow because it examines all function calls in the codebase.

## Risk

Any `async` function that indirectly calls `AsyncioManager.run()` will raise an error. Since we don't have a lot of `async` functions, I grepped for `async def` and examined everything (skipping the ones in `wandb/sdk/launch/` which do not depend on the code I'm changing).

I'm hoping test coverage is enough to detect anything I've missed.

## Testing

```python
from tqdm import trange
import wandb

with wandb.init() as run:
    run.define_metric("*", summary="none")

    for i in trange(100_000):
        run.log({"x": 1, "y": 2, "z": 3, "a": 4})
```

Before this PR: the progress bar moves quickly and appears to pause every couple seconds. After it reaches the end, there is a long pause before wandb's status messages get printed because the OperationStats requests are behind a long backlog of History requests.

After this PR: the progress bar moves slower but at a constant pace. Status messages show up almost immediately once it reaches the end.