fix: try to avoid conflicting run directory names (#10511)
A run's directory name depends on the run's start time and ID and has the format `[offline-]run-YYYYMMDD-hhmmss-<ID>`. The user can set the ID to anything, so it is entirely possible to create two runs with the same ID within the same second. When this happens, the behavior is undefined: some files that get created might belong to the first run instance, some to the other, the transaction log may be corrupted, and some code may just error out.

A partial workaround is for `wandb.init()` to check if the run directory already exists and add a suffix if it does. A conflict can still happen if the runs are made in parallel, like by two scripts.