refactor: rename StreamRun to RunHandle (#10538)
Renames `StreamRun` to `RunHandle`, gives it a more solid definition, and moves it to its own package.

This is a proper way of receiving the run's Entity and Project instead of relying on `Settings` being mutable. It's safer too: if the values haven't been initialized yet, code can error out instead of using potentially incorrect values.

Moving it to its own package is necessary so that the `runfiles` and `monitor` packages can reference it in a later PR. Even though it's closely related to `runupserter`​ at the moment, I put it in its own package to enable a future refactor: anything that uses `RunHandle` getters should instead be initialized as part of `RunHandle.Init()` and receive the objects it relies on as explicit parameters. This will reverse dependencies; e.g. we'll start with `runfiles => {runhandle, runupserter}`​ and end with `runhandle => {runfiles, runupserter}`​.