refactor: initialize SettingsStatic with a dict instead of a proto (#10593)
Instead of converting to a proto and then back to a dict, just pass a dict directly to `SettingsStatic`.

I will be removing `x_files_dir` from the proto but keeping it on the Python `Settings` object for legacy code.

Found all uses by searching for "SettingsStatic(". This class is not part of the public interface (it's not exported in `wandb/__init__.py`).

The `importers/internal.py::make_settings()`​function appears to not be tested, but I removed all fields that you aren't allowed to set explicitly on `Settings`​ (Pydantic raises an error). There were also some tests in `test_job_builder.py`​ that set `_jupyter`​ which isn't settable. The `SettingsStatic._proto_to_dict()`​ method used to silently ignore these fields.