feat: flow control (#10073)
Implements a mechanism to allow the SDK to off-load incoming data to disk when it is busy.

The new FlowControl component sits between the Writer and Sender and acts like an infinitely buffered channel for records that were written to the transaction log. Incoming records are buffered in-memory up to a limit, after which new records are discarded and re-read from the transaction log when needed.

For unsaved records, such as requests, the FlowControl component acts as a regular finitely-buffered channel. If it accumulates too many requests, it begins to block for all incoming records to create backpressure.

The key piece is the FlowControlBuffer which interleaves saved and unsaved records, preserving order but reducing memory usage for saved records.

NOTE: As of this PR, the FileStream component can still use an unbounded amount of memory. After this PR, it can be made to block upon reaching a limit.