fix: don't print empty status lines when output device is not a terminal (#10597)
When `wandb` is used in a script whose output is redirected to a file or a so-called "dumb" terminal, the progress updates it prints during `wandb.init()`, `run.finish()` and some other functions are simplified to avoid dynamic text with ANSI cursor motions and to instead print a single line each time the combined status message changes.

The code was incorrectly printing a blank line when the status was blank (meaning wandb-core didn't report any ongoing operations). This was noticed because it caused a test to flake in `test_offline_sync_beta.py`.