refactor: make FileStream.StreamUpdate() block when too much data is accumulated (#10454)
~~Before merging this, need to extract run-stop logic from the Sender or else we stop polling when FileStream is backed up.~~

Adds backpressure to  `FileStream`, making it stop accepting updates once it has accumulated a request of the maximum size (for testing, this is controllable using the `x_file_stream_max_bytes` setting). This causes the Sender to block and FlowControl (added in #10073) to kick in.

Can be tested with a script like:

```
from tqdm import trange
import wandb

settings = wandb.Settings(x_file_stream_max_bytes=10000)
with wandb.init(settings=settings) as run:
    run.define_metric("*", summary="none")
    for i in trange(1000000):
        run.log({"x": 1, "y": 2, "z": 3, "w": 4})
```

The message `"filestream: waiting to send request of max size"` will appear in `debug-internal.core`, but the script will continue to loop at the same speed indicating that FlowControl is active (or the socket buffer is big).