refactor: subtask error message indent logic in progress.py (#10718)
I recently added an `_INDENT`​ constant to this file and was tempted to use it in the `op.error_status` block, but the "indent" there is for alignment and not the same thing as what `_INDENT` represents. I also noticed it would not align properly in terminals that don't support Unicode, so I refactored it slightly.