chore: fix test_xgboost failure (#10727)
Fixes `test_xgboost.py` which is failing because of the recent `xgboost` release (3.1.0) which appears to have changed the number of features it loads from the Wine dataset from 11 to 13.

This test is poorly written, but really fixing it would take a lot of time and possibly require refactoring the integration as well.

I removed `use_label_encoder=False` because it was emitting a warning about being an unused argument.