refactor: runsummary Updates object (#10729)
Creates `runsummary.Updates` as the Go counterpart to `SummaryRecord`, splitting out this logic from `runsummary.RunSummary`.

This will allow passing updates instead of the full serialized summary to FileStream, which we're doing to remove the summary debouncing logic in `sender.go`.

I kept the `SummaryItem` proto in the return values of `UpdateSummaries` and `ToRecords` since that's what `handler.go` needs. It wouldn't make sense to convert to `Updates` and then back to `SummaryItem`.