refactor: move FileStream state into collector goroutine (#10733)
Moves FileStream state from the TransmitLoop goroutine to the CollectLoop goroutine.

The CollectLoop goroutine needs to estimate the size of the next request, and since the state is necessary for constructing a request, it needs to live in CollectLoop. This was unnecessary before because the state did not affect the size of a request (it just contained line offsets), but we'll need it to put the run summary into the state.

To avoid serializing the full JSON request every time the CollectLoop receives an update like

```go
for {
	select {
	// Running GetJSON on every iteration is expensive!
	case transmissions <- reader.GetJSON(state):
		... // start a new buffer
	case request := <-requests:
		... // incorporate into the buffer
	}
}
```

the new `TransmitChan` enables a semaphore-like pattern:

```go
for {
	select {
	// Here, GetJSON only runs if the TransmitLoop is ready.
	case transmissions := <-output.PreparePush():
		transmissions <- reader.GetJSON(state)
		...
	case request := <-requests:
		...
	}
}
```