fix: add nowait param to publish_output() for asyncio safety (#10728)
Adds a `nowait` parameter to `publish_output` and `publish_output_raw` to allow them to be called from the asyncio thread. When set to true, `AsyncioManager`'s `run_soon()` method is used instead of `run()` to submit the task to send the record over the socket, which makes it non-blocking even if the socket buffer is full.

This PR also improves the very outdated documentation for the `InterfaceBase` class hierarchy.

Before this PR, console output from wandb async functions (running in the AsyncioManager) was not captured and raised an exception that is logged and suppressed. One example is when the `asyncio` Socket code logs the message `socket.send() raised an exception.` if the underlying socket disconnects unexpectedly (like if `wandb-core` dies).