refactor: completely replace FileStreamRequestReader (#10765)
Replaces `FileStreamRequestReader` by rolling it into `FileStreamState` and significantly simplifying.

These were separate because the request buffer used to live in a different goroutine from the `FileStreamState`. They can be merged now that they're both in the `CollectLoop` goroutine.

The implementation is simpler: The `FileStreamRequestReader` used a "prepare then execute" pattern that selected a portion of the `FileStreamRequest` and separately split out a `(data_to_upload, remaining_data)` pair, requiring a concrete representation of the selection. That representation had to be performant because the "prepare" step (`NewRequestReader`) was used frequently, so it was essentially a count of history/events/console lines.

`FileStreamState` is simpler because it can drop data from `FileStreamRequest`, insert it into `FileStreamRequestJSON` and update the state all in one go.