refactor: improve mock_wandb_log assertion messages (#10877)
Replaces `assert mock_wandb_log.logged("message")` by `mock_wandb_log.assert_logged("message")`.

The new assertion methods print the logged messages when they don't contain the expected message.

There's no way to write `assert not mock_wandb_log.logged(...)` anymore, but that was a bad idea since it can't test that a different, similar message was not logged. The only example of this was in the weave integration tests, which I updated.