refactor: simplify Api._load_api_key (#10927)
Updates `_login` to make it usable in `Api._load_api_key` and fixes a bug in `_load_api_key`.

Pulls out the `_noop`, `_offline` and `wandb.run` checks into the higher level `wandb.login()` (which is also called by the `wandb login` CLI). The `_noop` and `_offline` checks are already present in `wandb.init()` and the `wandb.run` check is not relevant there. The fact that `Api._load_api_key` doesn't work if the `identity_token_file` setting is set is the same as it was before.

`_login` is called by `login()` (which is called by the `wandb login` CLI), `wandb.init()`, Sweeps, and now `Api._load_api_key()`. The changes above shouldn't matter for Sweeps.

I recently refactored `_load_api_key()` to use the new `auth` methods in PR #10893 but in doing so introduced a subtle bug: if `Api._load_api_key` prompts for a key, that key is not saved in the session, and later calls to `Api()` will prompt again. I will add an `auth` method for this soon, but due to other obstacles, using `_login` is the cleaner refactoring path.